+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [2.0a4] - Unreleased
 
+### Fixed
+
+- Fixed DoFSpace basis pseudoinverse calculation
+
 ### Added
 
 - Added gradients of correlations functionality in C++, Python functions